web/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
